add validateDate.js and validateDate.test.js components to reuse.  add date-fns package for date validation test.
