+++ f2/package.json
@@ -16,6 +16,7 @@
     "chroma-js": "^2.1.0",
     "clamdjs": "^1.0.2",
     "core-js": "^3.6.4",
+    "date-fns": "^2.12.0",
     "dotenv": "^8.2.0",
     "email-addresses": "^3.1.0",
     "emotion-theming": "^10.0.27",
